decompress: sanity check on K_msbs

+++ src/lib/jp2/grok.h
@@ -1010,7 +1010,7 @@ typedef struct _grk_plugin_code_block
 	uint32_t numPix;
 	uint8_t* compressedData;
 	uint32_t compressedDataLength;
-	size_t numBitPlanes;
+	uint8_t numBitPlanes;
 	size_t numPasses;
 	grk_plugin_pass passes[67];
 	unsigned int sortedIndex;
